+++ src/test/scala/edu/ie3/util/scala/collection/mutable/CountingMapSpec.scala
@@ -4,7 +4,7 @@
  * Research group Distribution grid planning and operation
  */
 
-package edu.ie3.util.scala
+package edu.ie3.util.scala.collection.mutable
 
 import edu.ie3.simona.test.common.UnitSpec
